QAChatBot/ingest.py
from langchain_huggingface import HuggingFaceEmbeddings
from langchain_community.vectorstores import FAISS
from langchain_community.document_loaders import PyPDFLoader, DirectoryLoader, JSONLoader
from langchain.text_splitter import RecursiveCharacterTextSplitter 
import requests


DATA_PATH = 'Data/'
DB_FAISS_PATH = 'LawDocbase/db_faiss'  #" //upload/db(botid)/db/db_fiss"

print("DB_FAISS_PATH = ", DB_FAISS_PATH)

# Create vector database
def create_vector_db():
    loader = DirectoryLoader(DATA_PATH,
                             glob='*.pdf',
                             loader_cls=PyPDFLoader
                            #  glob='*.json',
                            #  loader_cls=JSONLoader, loader_kwargs={'jq_schema':'.title_identification','text_content':False}
                             )

    documents = loader.load()
    text_splitter = RecursiveCharacterTextSplitter(chunk_size=500,
                                                   chunk_overlap=50)
    texts = text_splitter.split_documents(documents)

    embeddings = HuggingFaceEmbeddings(model_name='sentence-transformers/all-mpnet-base-v2',
                                       model_kwargs={'device': 'cpu'})

    db = FAISS.from_documents(texts, embeddings)
    db.save_local(DB_FAISS_PATH)

if __name__ == "__main__":
    create_vector_db()

